Able to now load any extensions
