+++ src/Extensions/Text/ChatManager.py
@@ -1,7 +1,7 @@
 from discord.ext import commands
 
 
-class ChatManager(commands.bot):
+class ChatManager(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
